requirements.txt
PyYAML==5.4.1
requests==2.26.0
tabulate==0.8.9